Introduce Cost Sensor
